+++ immutable/package.json
@@ -1,8 +1,11 @@
 {
   "name": "swr-immutable",
   "version": "0.0.1",
+  "main": "../dist/immutable.js",
+  "module": "../dist/immutable.esm.js",
   "peerDependencies": {
-    "swr": "^1.0.0",
-    "react": "^16.11.0 || ^17.0.0"
+    "swr": "*",
+    "react": "*",
+    "dequal/lite": "*"
   }
 }
