+++ infinite/package.json
@@ -4,7 +4,8 @@
   "main": "../dist/infinite.js",
   "module": "../dist/infinite.esm.js",
   "peerDependencies": {
-    "swr": "^1.0.0",
-    "react": "^16.11.0 || ^17.0.0"
+    "swr": "*",
+    "react": "*",
+    "dequal/lite": "*"
   }
 }
