+++ package.json
@@ -23,9 +23,11 @@
   "homepage": "https://swr.vercel.app",
   "license": "MIT",
   "scripts": {
-    "build": "bunchee src/index.ts -m --no-sourcemap",
+    "build": "yarn build:core && yarn build:infinite && yarn build:immutable",
     "watch": "bunchee src/index.ts --watch",
+    "build:core": "bunchee src/index.ts -m --no-sourcemap",
     "build:infinite": "bunchee index.ts --cwd infinite -m --no-sourcemap",
+    "build:immutable": "bunchee index.ts --cwd immutable -m --no-sourcemap",
     "prepublishOnly": "rm -rf dist && yarn build",
     "types:check": "tsc --noEmit",
     "format": "prettier --write \"{src,test,examples}/**/*.{ts,tsx}\"",
