+++ src/utils/normalize-args.ts
@@ -1,0 +1,12 @@
+import { Key, Fetcher, SWRConfiguration } from '../types'
+
+export function normalize<KeyType = Key, Data = any>(
+  args:
+    | readonly [KeyType]
+    | readonly [KeyType, SWRConfiguration | undefined]
+    | readonly [KeyType, Fetcher<Data> | null, SWRConfiguration | undefined]
+): [KeyType, Fetcher<Data> | null, Partial<SWRConfiguration<Data>>] {
+  return typeof args[1] === 'function'
+    ? [args[0], args[1], args[2] || {}]
+    : [args[0], null, (typeof args[1] === 'object' ? args[1] : args[2]) || {}]
+}
